+++ packages/local-signing-manager/src/lib/local-signing-manager.ts
@@ -8,9 +8,13 @@ import {
 } from '@polkadot/types/types';
 import { hexToU8a, u8aToHex } from '@polkadot/util';
 import { cryptoWaitReady, mnemonicGenerate } from '@polkadot/util-crypto';
-import { PolkadotSigner, SigningManager } from '@polymeshassociation/signing-manager-types';
+import {
+  PolkadotSigner,
+  signedExtensions,
+  SigningManager,
+} from '@polymeshassociation/signing-manager-types';
 
-import { PrivateKey } from '../types';
+import { KeyRingType, PrivateKey } from '../types';
 
 /**
  * Manages signing payloads with a set of pre-loaded accounts in a Keyring
@@ -28,11 +32,9 @@ export class KeyringSigner implements PolkadotSigner {
    */
   public async signPayload(payload: SignerPayloadJSON): Promise<SignerResult> {
     const { registry } = this;
-    const { address, signedExtensions, version } = payload;
+    const { address, version } = payload;
 
     const pair = this.getPair(address);
-
-    registry.setSignedExtensions(signedExtensions);
 
     const signablePayload = registry.createType('ExtrinsicPayload', payload, {
       version,
@@ -101,10 +103,15 @@ export class LocalSigningManager implements SigningManager {
    *
    * @param args.accounts - array of private keys
    */
-  public static async create(args: { accounts: PrivateKey[] }): Promise<LocalSigningManager> {
+  public static async create(args: {
+    accounts: PrivateKey[];
+    type?: KeyRingType;
+  }): Promise<LocalSigningManager> {
     await cryptoWaitReady();
 
-    return new LocalSigningManager(args.accounts);
+    const { accounts, type } = args;
+
+    return new LocalSigningManager(accounts, type);
   }
 
   /**
@@ -119,12 +126,15 @@ export class LocalSigningManager implements SigningManager {
   /**
    * @hidden
    */
-  private constructor(accounts: PrivateKey[]) {
+  private constructor(accounts: PrivateKey[], type?: KeyRingType) {
     this.keyring = new Keyring({
-      type: 'sr25519',
+      type: type || 'sr25519',
     });
 
-    this.externalSigner = new KeyringSigner(this.keyring, new TypeRegistry());
+    const registry = new TypeRegistry();
+    registry.setSignedExtensions(signedExtensions);
+
+    this.externalSigner = new KeyringSigner(this.keyring, registry);
 
     accounts.forEach(account => {
       this._addAccount(account);